fetch new openvpn-monitor release
The 1.0.0 release supports GeoIP2 API.
Previously used release-candidate blob doesn't.
The the release requires presence of `openvpn-monitor.conf` file during installation. This heats up the division of docker's `confd` vs. openvpn-monitor's `openvpn-monitor.conf` for configuration. I'll stick with the latter for now, as editing templates for `confd` is a sorcery for me.